pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.29.0",
    "sift-stack-py>=0.8.2",
    "streamlit>=1.48.0",
]
